refactor: separate module for provider handler
